ThomasB : Changement pour restriction idUser, Création Supression et …

+++ controller/post_controller.js
@@ -1,5 +1,5 @@
-const postModel = require("../model/post_model");
-const Post = require("../model/user_model");
+const Post = require("../model/post_model");
+const Commentaire = require("../model/commentaire_model");
 require("dotenv").config();
 
 // permet de creer un post
@@ -7,10 +7,10 @@ exports.create = (req, res, next) => {
     if (req.body.titlePost === "" || req.body.contentPost === "") {
       return res.status(400).json({ error: "Veuillez remplir tous les Champs pour votre POST SVP ! " });
     }
-    postModel.create({
+    Post.create({
       titlePost: req.body.titlePost,
       contentPost: req.body.contentPost,
-      user: req.token.email
+      user: req.token.id
     })
       .then(() => res.status(201).json({ message: "Post enregistré !" }))
       .catch((error) => res.status(400).json({ error }));
@@ -19,12 +19,16 @@ exports.create = (req, res, next) => {
 exports.delete = async (req, res, next) => {
   idPost = req.params;
   const dataPost = await Post.findById(idPost);
-  // if (req.token.id === dataPost.idUser){
-  await Post.deleteOne({_id:idPost});
-  res.status(201).json({message:"Post supprimé"})
-  // }else{
-  //   res.status(403).json({ message:"Vous n'avez pas les droits pour supprimer ce post"})
-  // }
+  if (req.token._id === dataPost.idUser){
+    if(dataPost == null){
+      res.status(404).json({ message:"Le post que vous shouaitez suprimmer n'existe pas"})
+    }else{
+      await Post.deleteOne({_id:idPost});
+      res.status(201).json({message:"Post supprimé"});
+    }
+  }else{
+    res.status(403).json({ message:"Vous n'avez pas les droits pour supprimer ce post"})
+  }
 }
 
 exports.getAll = async (req, res, next) => {
@@ -36,4 +40,11 @@ exports.getOne = async (req, res, next) => {
   idPost = req.params;
   const dataPost = await Post.findById(idPost);
   res.status(200).json(dataPost);
+}
+
+exports.getOneAndCommentaire = async (req, res, next) => {
+  idPost = req.params;
+  const dataPost = await Post.findById(idPost);
+  const dataCommentaire = await Commentaire.find({idPost : idPost});
+  res.status(200).json({post:dataPost,commentaire:dataCommentaire});
 }
